OMG it just randomly works now
WASI was throwing error even on sys, saying that "environment-cli" was getting the wrong import on something.

It was happening on worlds without imports, with imports it was fine. Even managed to "recreate" it once, but after proper minimal viable example reproduction, it started working suddenly.

Maybe switching to cargo component version 0.5 fixed the problem?